Refactor coloring into separate hook
